Rename setup directory to install. (#1551)
* rename setup directory to install

* tweak a few more files
