+++ .travis.yml
@@ -23,15 +23,15 @@ services:
   - docker
 
 before_install:
-  - ./setup/scripts/docker_setup_forseti.sh
+  - ./install/scripts/docker_setup_forseti.sh
 
 install:
   -
 
 script:
-  - ./setup/scripts/docker_install_forseti.sh
-  - ./setup/scripts/docker_unittest_forseti.sh
-  - ./setup/scripts/docker_pylint_forseti.sh
+  - ./install/scripts/docker_install_forseti.sh
+  - ./install/scripts/docker_unittest_forseti.sh
+  - ./install/scripts/docker_pylint_forseti.sh
 
 after_success:
   # Collect codecov results.
