+++ install/docker/base
@@ -4,7 +4,7 @@
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
 #
-#     http://www.apache.org/licenses/LICENSE-2.0
+#    http://www.apache.org/licenses/LICENSE-2.0
 #
 # Unless required by applicable law or agreed to in writing, software
 # distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,12 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-""" Forseti Installer.
+FROM ubuntu:18.04
 
-A stub to call setup/gcp/run_forseti_installer.py which installs into GCP.
-"""
-from gcp import run_forseti_installer
+# Expose our source so we can read in dependencies.
+ADD . /forseti-security/
+WORKDIR /forseti-security/
 
+# Install host dependencies.
+RUN apt-get update -qq 1> /dev/null
+RUN apt-get install -qq -y $(cat install/dependencies/apt_packages.txt | grep -v "#" | xargs) 1> /dev/null
+RUN rm -rf /var/lib/apt/lists/*
 
-if __name__ == '__main__':
-    run_forseti_installer.run()
+# Install the CloudSDK for `gcloud`.
+RUN curl -sSL https://sdk.cloud.google.com 1> /dev/null | bash
