+++ install/gcp/installer/util/files.py
